third_party/blink/web_tests/fast/url/scheme-expected.txt

Canonicalization of URL schemes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS canonicalize('http://example.com/') is 'http://example.com/'
PASS canonicalize('HTTP://example.com/') is 'http://example.com/'
PASS successfullyParsed is true

TEST COMPLETE
